#54 - Update how StorageOperation paths are generated
From now on we give the full path to the StorageOperation. That means for
replicating a bag you get up to the root of the bag (/path/to/bag/), and
for a token store you get the token store (/path/to/tokens). This has a caveat
that the rsyncs for bags will need to end with a slash so that the directory is
not recreated, which we will look into next.