Avoid concurrent session usage
This isn't the best implementation and it would be better to either:
create the session prior and pass it in (requires knowledge in lower
modules) or allocating a session for each thread (not sure how this
would be cleaned up - maybe time based). The naive impl is fine just
things to consider for the future.